fed/architecture: Correctly add image
rst2pdf considers named substitution as inline image and just expects the image
to be of same height as the text.

Adding width to the image tells rst2pdf that this image should be as wide as the
page, otherwise this image was looking very small and ugly. Now it looks a
little torn, maybe we need higher quality image from swimlanes.io